Terminada a UI  y pruebas con la DB y sus datos

+++ src/uiMain/UserInterface.java
@@ -16,10 +16,9 @@ public class UserInterface {
 	public static void main(String[] args) {
 
 		Reader.Read();
-		
+
 		Scanner scan = new Scanner(System.in);
-		
-		
+
 		int metodoPago;
 		int edad;
 		String debCred = "";
@@ -32,71 +31,66 @@ public class UserInterface {
 
 		List<Producto> productos = new ArrayList<Producto>();
 		List<Integer> cantidades = new ArrayList<Integer>();
-		
-		System.out.print(Mensajes.edad());
+
+		System.out.print(Mensajes.edad);
 		edad = scan.nextInt();
-		System.out.print(Mensajes.debCred());
+		System.out.print(Mensajes.debCred);
 		metodoPago = scan.nextInt();
-		
-		if(metodoPago == 1) { debCred = "Debito"; }
-		else if(metodoPago == 2) { debCred = "Credito"; }
-		
-		cliente = new Cliente(edad, debCred);
-		
-		System.out.print(Mensajes.bienvenida());
-		System.out.println(Mensajes.seccionar());
+
+		cliente = MenuUsuario.crearCliente(metodoPago, edad);
+
+		System.out.print(Mensajes.bienvenida);
+		System.out.println(Mensajes.seccionar);
 
 		while (!exit) {
 
-			System.out.println(Mensajes.selInicio());
-			System.out.println(Mensajes.seccionar());
-			System.out.print(Mensajes.opcion());
-			
+			System.out.println(Mensajes.selInicio);
+			System.out.println(Mensajes.seccionar);
+			System.out.print(Mensajes.opcion);
+
 			option = scan.nextInt();
-			
-			
+
 			switch (option) {
 			case 1:
-				
+
 				boolean menuUsuario = false;
 				while (!menuUsuario) {
-					
-				System.out.println(Mensajes.seccionar());
-				System.out.println(Mensajes.selUsuario());
-				System.out.println(Mensajes.seccionar());
-				System.out.print(Mensajes.opcion());
 
-				option = scan.nextInt();
+					System.out.println(Mensajes.seccionar);
+					System.out.println(Mensajes.menuCliente);
+					System.out.println(Mensajes.selUsuario);
+					System.out.println(Mensajes.seccionar);
+					System.out.print(Mensajes.opcion);
 
-				switch (option) {
-				case 1:
-					
-					System.out.println(cliente.getNivel_afiliacion());
-
-					break;
-
-				case 2:
-					
-					System.out.println(Mensajes.afiliacion());
-					System.out.print(Mensajes.opcion());
 					option = scan.nextInt();
-					
-					if (option == 1) { cliente.aumentarAfiliacion(); }
-					else if(option == 2) { cliente.disminuirAfiliacion(); }
-					break;
 
-				case 3:
-					
-					if (cliente.getFacturas().size() == 0) { System.out.println("No tienes facturas para mostrar."); }
-					else if(cliente.getFacturas().size() > 0) { System.out.println(cliente.getFacturas().size()); } //Mostrar con el metodo toString
-					
-					break;
+					switch (option) {
+					case 1:
 
-				default:
-					menuUsuario = true;
-					break;
+						System.out.println(Mensajes.mensajeNivelAfiliacion + cliente.getNivel_afiliacion());
+
+						break;
+
+					case 2:
+
+						System.out.println(Mensajes.afiliacion);
+						System.out.print(Mensajes.opcion);
+						option = scan.nextInt();
+
+						MenuUsuario.modifAfiliacion(option, cliente);
+						break;
+
+					case 3:
+
+						MenuUsuario.facturasCliente(cliente);
+
+						break;
+
+					default:
+						menuUsuario = true;
+						break;
+					}
 				}
-			}
 				break;
 
 			case 2:
@@ -107,25 +101,26 @@ public class UserInterface {
 					int cantidad = -1;
 					int index = 0;
 
-					System.out.println(Mensajes.seccionar());
-					System.out.println("Menu tienda.");
+					System.out.println(Mensajes.seccionar);
+					System.out.println(Mensajes.menuTienda);
 
-					System.out.println(Mensajes.selTienda());
-					System.out.println(Mensajes.seccionar());
-					System.out.print(Mensajes.opcion());
+					System.out.println(Mensajes.selTienda);
+					System.out.println(Mensajes.seccionar);
+					System.out.print(Mensajes.opcion);
 
 					option = scan.nextInt();
 
 					switch (option) {
 
-
 					case 1:
 
+						System.out.println(Mensajes.seccionar);
 						MenuProducto.mostrarProducto();
+						System.out.println(Mensajes.seccionar);
 
-						System.out.print("Seleccione un id de producto: ");
+						System.out.print(Mensajes.selID);
 						option = scan.nextInt();
-						System.out.print("Cantidad: ");
+						System.out.print(Mensajes.cantidad);
 						cantidad = scan.nextInt();
 
 						MenuTienda.validarProducto(option, cantidad, productos, cantidades);
@@ -133,29 +128,31 @@ public class UserInterface {
 						break;
 
 					case 2:
+
+						System.out.println(Mensajes.seccionar);
 						MenuTienda.mostrarListaProductos(productos, cantidades);
+						System.out.println(Mensajes.seccionar);
+						if (productos.size() > 0) {
+							System.out.print(Mensajes.selID);
+							idProducto = scan.nextInt();
+							System.out.print(Mensajes.cantidad);
+							cantidadProducto = scan.nextInt();
 
-						System.out.print("ingrese el id del producto: ");
-						idProducto = scan.nextInt();
-						System.out.print("Cantidad a eliminar: ");
-						cantidadProducto = scan.nextInt();
+							MenuTienda.eliminarProducto(productos, cantidades, idProducto, cantidadProducto);
+						}
 
-						MenuTienda.eliminarProducto(productos, cantidades, idProducto, cantidadProducto);
 						break;
 
 					case 3:
+						System.out.println(Mensajes.seccionar);
 						MenuTienda.mostrarListaProductos(productos, cantidades);
+
 						break;
 
 					case 4:
-						totalVenta = 0;
-						for (int i = 0; i < productos.size(); i++) {
-							totalVenta += productos.get(i).getPrecio() * cantidades.get(i);
-						}
+						System.out.println(Mensajes.seccionar);
+						MenuTienda.pagar(productos, cantidades, cliente);
 
-						new Factura(productos, cantidades, totalVenta, cliente);
-						productos.clear();
-						cantidades.clear();
 						break;
 
 					default:
@@ -171,18 +168,19 @@ public class UserInterface {
 
 				boolean menuProductos = false;
 				while (!menuProductos) {
-					
-					System.out.println(Mensajes.seccionar());
-					System.out.println("Menu Productos.");
-					System.out.println(Mensajes.selProductos());
-					System.out.println(Mensajes.seccionar());
-					System.out.print(Mensajes.opcion());
+
+					System.out.println(Mensajes.seccionar);
+					System.out.println(Mensajes.menuProducto);
+					System.out.println(Mensajes.selProductos);
+					System.out.println(Mensajes.seccionar);
+					System.out.print(Mensajes.opcion);
 					option = scan.nextInt();
 
 					switch (option) {
 					case 1:
 
 						MenuProducto.crearProducto(1, scan);
+
 						break;
 
 					case 2:
@@ -194,22 +192,26 @@ public class UserInterface {
 
 						MenuProducto.crearProducto(3, scan);
 						break;
-					
+
 					case 4:
+						System.out.println(Mensajes.seccionar);
 						MenuProducto.mostrarProducto();
-						System.out.println("1. Agregar. \n2. Eliminar");
+						System.out.println(Mensajes.seccionar);
+
+						System.out.println("1. Agregar. / 2. Eliminar");
+						System.out.print(Mensajes.opcion);
 						option = scan.nextInt();
-						System.out.print("Ingrese el id del producto: ");
+						System.out.print(Mensajes.selID);
 						idProducto = scan.nextInt();
-						System.out.print("Cantidad de invenatiro a agregar: ");
+						System.out.print("Cantidad de inventario a Agregar/Quitar: ");
 						cantidadProducto = scan.nextInt();
-						
+
 						MenuProducto.actualizarInventario(option, idProducto, cantidadProducto);
-						
+
 						break;
 
 					case 5:
-
+						System.out.println(Mensajes.seccionar);
 						MenuProducto.mostrarProducto();
 						break;
 
@@ -219,26 +221,42 @@ public class UserInterface {
 					}
 				}
 				break;
-				
-			case 4:
 
-				System.out.println(Factura.facturas.size()); //metodo toString
-				System.out.println(Mensajes.seccionar());
+			case 4:
+				System.out.println(Mensajes.seccionar);
+				System.out.println(Factura.facturas.size()); // metodo toString
+				System.out.println(Mensajes.seccionar);
 				break;
-				
+
 			case 5:
+				System.out.println(Mensajes.seccionar);
+				System.out.println("El producto mas vendido es: " + Factura.masVendido());
+				System.out.println(Mensajes.seccionar);
+
+				break;
+
+			case 6:
+				System.out.println(Mensajes.seccionar);
+				System.out.println(Factura.cantProductVentDebCred());
+				System.out.println(Mensajes.seccionar);
+				break;
+
+			case 7:
+				System.out.println(Mensajes.seccionar);
+				System.out.println(Factura.ventasTotales());
+				System.out.println(Mensajes.seccionar);
+				break;
+
+			case 8:
+				System.out.println(Mensajes.seccionar);
+				System.out.println(Factura.gananciaNeta());
+				System.out.println(Mensajes.seccionar);
+				break;
+			case 9:
+				System.out.println(Mensajes.seccionar);
 				System.out.println("Adios.");
 				Writer.Write();
 				exit = true;
-				break;
-				
-			case 6:
-				System.out.println("El producto mas vendido es: " + Factura.masVendido());
-				System.out.println(Mensajes.seccionar());
-				break;
-				
-			case 7:
-				//System.out.println(Factura.cantProductVentDebCred());
 				break;
 			default:
 				System.out.println("Opcion incorrecta. ");
@@ -248,8 +266,14 @@ public class UserInterface {
 	}
 
 	// Functions
-	
 
-	
-	
+	// totalVenta = 0;
+	// for (int i = 0; i < productos.size(); i++) {
+	// totalVenta += productos.get(i).getPrecio() * cantidades.get(i);
+	// }
+
+	// new Factura(productos, cantidades, totalVenta, cliente);
+	// productos.clear();
+	// cantidades.clear();
+
 }
